+++ projects/ngx-mat-lib/src/lib/xmat-validators/index.ts
@@ -3,3 +3,4 @@ export * from "./xmat-min-list-selection.validation";
 export * from "./xmat-valid-list-selection.validation";
 export * from "./xmat-values-match.validation";
 export * from "./xmat-fill-at-least.validation";
+export { xmatMinTimeValidation, xmatMaxTimeValidation } from "./xmat-time-input.validator";