+++ src/app/views/xmat-accordion/xmat-accordion-examples.component.ts
@@ -9,6 +9,7 @@ export class XmatAccordionExamplesComponent {
 
     accordionOpened: boolean = !0;
     accordionDisabled: boolean = !0;
+    accordionBOpened: boolean = !0;
 
     constructor() {